requirements.txt
aiogram==3.22.0
openai==1.47.0
python-dotenv==1.0.1
pandas==2.2.2
gspread==6.1.2
oauth2client==4.1.3
openpyxl==3.1.5
aiohttp==3.12.15
certifi==2025.10.5
